moved back the v1 APIs to apis/ (#285)
* moved back the v1 APIs to apis/
because it did not matter where they were since operator-sdk would
always generate the CRD YAML regardless.


* removed unnecessary CRDs from base CSV
