+++ apis/acme.cert-manager/v1/zz_generated.deepcopy.go
@@ -22,7 +22,7 @@ limitations under the License.
 package v1
 
 import (
-	meta_cert_managerv1 "github.com/ibm/ibm-cert-manager-operator/v1apis/meta.cert-manager/v1"
+	meta_cert_managerv1 "github.com/ibm/ibm-cert-manager-operator/apis/meta.cert-manager/v1"
 	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 )
