+++ apis/cert-manager/v1/issuer_types.go
@@ -20,8 +20,8 @@ package v1
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	cmacme "github.com/ibm/ibm-cert-manager-operator/v1apis/acme.cert-manager/v1"
-	cmmeta "github.com/ibm/ibm-cert-manager-operator/v1apis/meta.cert-manager/v1"
+	cmacme "github.com/ibm/ibm-cert-manager-operator/apis/acme.cert-manager/v1"
+	cmmeta "github.com/ibm/ibm-cert-manager-operator/apis/meta.cert-manager/v1"
 )
 
 // +genclient
