+++ controllers/cert-manager/v1_add_label_controller.go
@@ -35,8 +35,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
+	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/apis/cert-manager/v1"
 	res "github.com/ibm/ibm-cert-manager-operator/controllers/resources"
-	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/v1apis/cert-manager/v1"
 )
 
 // V1AddLabelReconciler reconciles a Certificate object
