+++ controllers/cert-manager/v1alpha1_add_label_controller.go
@@ -30,9 +30,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
+	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/apis/cert-manager/v1"
 	certmanagerv1alpha1 "github.com/ibm/ibm-cert-manager-operator/apis/certmanager/v1alpha1"
 	res "github.com/ibm/ibm-cert-manager-operator/controllers/resources"
-	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/v1apis/cert-manager/v1"
 )
 
 // V1Alpha1AddLabelReconciler reconciles a Certificate object
