+++ controllers/certmanager/issuer_controller.go
@@ -20,8 +20,8 @@ import (
 	"context"
 	"time"
 
+	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/apis/cert-manager/v1"
 	"github.com/ibm/ibm-cert-manager-operator/controllers/operator"
-	certmanagerv1 "github.com/ibm/ibm-cert-manager-operator/v1apis/cert-manager/v1"
 	"golang.org/x/mod/semver"
 	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
